3 eps modes
Mode 1 is the first implementation in version 5.0.: only param
eps_distance_gain is needed. In this mode, every time a position data is
recieved a estimation is calculated.

Mode 2 is the current EPS system, params eps_distance_gain and
eps_frequency es mandatory. In this mode a estimation is calculated
every time the clock reach the eps_frequency. No estimation is
calculated when a position data is recieved.

Mode 3 is a mix of mode 1 and mode 2: estimations are calculated every
time the timer reach the eps_frequency. The timer is reseted every time
a new position data is recieved.